Fix a couple of problems in how wsl session window titles were being handled:
1. the window title being set for a new wsl session contained more than
one colon and that was causing current path extraction to return an
invalid path value.

2. when a new session is started, the window title for a new wsl session
initially does not contain a valid path...due to temporary instability in
the value. this was causing a 'No such file or directory' error to
bleed out onto the console, an error that is actually of no consequence
...except that it made me go searching for "the problem". so, the code is
changed to issue a warning in this case instead.